Crate miden_crypto

Re-exports

pub use word::Word;

pub use word::WordError;

Modules

aead

AEAD (authenticated encryption with associated data) schemes.

dsa

Digital signature schemes supported by default in the Miden VM.

ecdh

ECDH (Elliptic Curve Diffie-Hellman) key agreement implementations.

field

Traits and utilities for working with the Goldilocks finite field (i.e., Felt).

hash

Cryptographic hash functions used by the Miden protocol.

ies

Integrated Encryption Scheme (IES) utilities.

merkle

Data structures related to Merkle trees based on Poseidon2 hash function.

parallel

Conditional parallel iteration primitives.

rand

Pseudo-random element generation.

stark

Foundational components for the STARK proving system based on Plonky3.

utils

Utilities used in this crate which can also be generally useful downstream.

word

A Word type used in the Miden protocol and associated utilities.

Macros

word

Construct a new Word from a hex value.

Structs

Felt

The prime field known as Goldilocks, defined as F_p where p = 2^64 - 2^32 + 1.

Enums

MapEntry

A view into a single entry in a map, which may either be vacant or occupied.

Constants

EMPTY_WORD

Array of field elements representing word of ZEROs in the Miden base field.

ONE

Field element representing ONE in the Miden base filed.

WORD_SIZE

Number of field elements in a word.

ZERO

Field element representing ZERO in the Miden base filed.

Traits

SequentialCommit

Defines how to compute a commitment to an object represented as a sequence of field elements.

Type Aliases

Map

An alias for a key-value map.

Set

An alias for a simple set.