miden_crypto::aead::xchacha

Struct SecretKey

pub struct SecretKey(/* private fields */);

A 256-bit secret key

Implementations

impl SecretKey

pub fn new() -> Self

Creates a new random secret key using the default random number generator

pub fn with_rng<R: RngCore + CryptoRng>(rng: &mut R) -> Self

Creates a new random secret key using the provided random number generator

pub fn encrypt_bytes( &self, data: &[u8], ) -> Result<EncryptedData, EncryptionError>

Encrypts and authenticates the provided data using this secret key and a random nonce

pub fn encrypt_bytes_with_associated_data( &self, data: &[u8], associated_data: &[u8], ) -> Result<EncryptedData, EncryptionError>

Encrypts the provided data and authenticates both the ciphertext as well as the provided associated data using this secret key and a random nonce

pub fn encrypt_bytes_with_nonce( &self, data: &[u8], associated_data: &[u8], nonce: Nonce, ) -> Result<EncryptedData, EncryptionError>

Encrypts the provided data using this secret key and a specified nonce

pub fn encrypt_elements( &self, data: &[Felt], ) -> Result<EncryptedData, EncryptionError>

Encrypts and authenticates the provided sequence of field elements using this secret key and a random nonce.

pub fn encrypt_elements_with_associated_data( &self, data: &[Felt], associated_data: &[Felt], ) -> Result<EncryptedData, EncryptionError>

Encrypts the provided sequence of field elements and authenticates both the ciphertext as well as the provided associated data using this secret key and a random nonce.

pub fn encrypt_elements_with_nonce( &self, data: &[Felt], associated_data: &[Felt], nonce: Nonce, ) -> Result<EncryptedData, EncryptionError>

Encrypts the provided sequence of field elements and authenticates both the ciphertext as well as the provided associated data using this secret key and the specified nonce.

pub fn decrypt_bytes( &self, encrypted_data: &EncryptedData, ) -> Result<Vec<u8>, EncryptionError>

Decrypts the provided encrypted data using this secret key.

Errors

Returns an error if decryption fails or if the underlying data was encrypted as elements rather than as bytes.

pub fn decrypt_bytes_with_associated_data( &self, encrypted_data: &EncryptedData, associated_data: &[u8], ) -> Result<Vec<u8>, EncryptionError>

Decrypts the provided encrypted data given some associated data using this secret key.

Errors

Returns an error if decryption fails or if the underlying data was encrypted as elements rather than as bytes.

pub fn decrypt_elements( &self, encrypted_data: &EncryptedData, ) -> Result<Vec<Felt>, EncryptionError>

Decrypts the provided encrypted data using this secret key.

Errors

Returns an error if decryption fails or if the underlying data was encrypted as bytes rather than as field elements.

pub fn decrypt_elements_with_associated_data( &self, encrypted_data: &EncryptedData, associated_data: &[Felt], ) -> Result<Vec<Felt>, EncryptionError>

Decrypts the provided encrypted data, given some associated data, using this secret key.

Errors

Returns an error if decryption fails or if the underlying data was encrypted as bytes rather than as field elements.