miden_crypto

Module field

Traits and utilities for working with the Goldilocks finite field (i.e., Felt).

Structs

BinomialExtensionField

Traits

BasedVectorSpace

A vector space V over F with a fixed basis. Fixing the basis allows elements of V to be converted to and from DIMENSION many elements of F which are interpreted as basis coefficients.

ExtensionField

A field EF which is also an algebra over a field F.

Field

A field F. This permits both modular fields ℤ/p along with their field extensions.

PrimeCharacteristicRing

A commutative ring, R, with prime characteristic, p.

PrimeField64

A prime field ℤ/p with order, p < 2^64.

QuotientMap

Implementation of the quotient map ℤ -> ℤ/p which sends an integer r to its conjugacy class [r].

TwoAdicField

A field which supplies information like the two-adicity of its multiplicative group, and methods for obtaining two-adic generators.

Functions

batch_inversion_allow_zeros

Parallel batch inversion using Montgomery’s trick, with zeros left unchanged.

batch_multiplicative_inverse

Batch multiplicative inverses with Montgomery’s trick This is Montgomery’s trick. At a high level, we invert the product of the given field elements, then derive the individual inverses from that via multiplication.