miden_crypto::hash

Trait HasherExt

pub trait HasherExt {
    type Digest;

    // Required method
    fn hash_iter<'a>(slices: impl Iterator<Item = &'a [u8]>) -> Self::Digest;
}

Extension trait for hashers to provide iterator-based hashing.

Required Associated Types

type Digest

The digest type produced by this hasher.

Required Methods

fn hash_iter<'a>(slices: impl Iterator<Item = &'a [u8]>) -> Self::Digest

Hashes an iterator of byte slices.

This method allows for more efficient hashing by avoiding the need to allocate a contiguous buffer when the input data is already available as discrete slices.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.