miden_crypto

Module merkle

Data structures related to Merkle trees based on Poseidon2 hash function.

Modules

mmr

Merkle Mountain Range (MMR) data structures.

smt

Sparse Merkle Tree (SMT) data structures.

store

Merkle store for efficiently storing multiple Merkle trees with common subtrees.

Structs

EmptySubtreeRoots

Contains precomputed roots of empty subtrees in a Merkle tree.

InnerNodeInfo

Representation of a node with two children used for iterating over containers.

MerklePath

A merkle path container, composed of a sequence of nodes of a Merkle tree.

MerkleProof

A container for a crate::Word value and its MerklePath opening.

MerkleTree

A fully-balanced binary Merkle tree (i.e., a tree where the number of leaves is a power of two).

NodeIndex

Address to an arbitrary node in a binary tree using level order form.

PartialMerkleTree

A partial Merkle tree with NodeIndex keys and 4-element Word leaf values. Partial Merkle Tree allows to create Merkle Tree by providing Merkle paths of different lengths.

RootPath

A container for a MerklePath and its crate::Word root.

SparseMerklePath

A different representation of MerklePath designed for memory efficiency for Merkle paths with empty nodes.

Enums

MerkleError

Functions

path_to_text

Utility to visualize a MerklePath in text.

tree_to_text

Utility to visualize a MerkleTree in text.