miden_crypto::merkle

Module mmr

Merkle Mountain Range (MMR) data structures.

Structs

Forest

A compact representation of trees in a forest. Used in the Merkle forest (MMR).

InOrderIndex

Index of nodes in a perfectly balanced binary tree based on an in-order tree walk.

Mmr

A fully materialized Merkle Mountain Range, with every tree in the forest and all their elements.

MmrDelta

Container for the update data of a super::PartialMmr

MmrPath

MmrPeaks

MmrProof

PartialMmr

Partially materialized Merkle Mountain Range (MMR), used to efficiently store and update the authentication paths for a subset of the elements in a full MMR.

Enums

MmrError