miden_crypto::merkle::mmr

Struct Mmr

pub struct Mmr { /* private fields */ }

A fully materialized Merkle Mountain Range, with every tree in the forest and all their elements.

Since this is a full representation of the MMR, elements are never removed and the MMR will grow roughly O(2n) in number of leaf elements.

Implementations

impl Mmr

pub fn new() -> Mmr

Constructor for an empty Mmr.

pub const fn forest(&self) -> Forest

Returns the MMR forest representation. See Forest.

pub fn open(&self, pos: usize) -> Result<MmrProof, MmrError>

Returns an MmrProof for the leaf at the specified position.

Note: The leaf position is the 0-indexed number corresponding to the order the leaves were added, this corresponds to the MMR size prior to adding the element. So the 1st element has position 0, the second position 1, and so on.

Errors

Returns an error if the specified leaf position is out of bounds for this MMR.

pub fn open_at(&self, pos: usize, forest: Forest) -> Result<MmrProof, MmrError>

Returns an MmrProof for the leaf at the specified position using the state of the MMR at the specified forest.

Note: The leaf position is the 0-indexed number corresponding to the order the leaves were added, this corresponds to the MMR size prior to adding the element. So the 1st element has position 0, the second position 1, and so on.

Errors

Returns an error if:

The specified leaf position is out of bounds for this MMR.

The specified forest value is not valid for this MMR.

pub fn get(&self, pos: usize) -> Result<Word, MmrError>

Returns the leaf value at position pos.

Note: The leaf position is the 0-indexed number corresponding to the order the leaves were added, this corresponds to the MMR size prior to adding the element. So the 1st element has position 0, the second position 1, and so on.

pub fn add(&mut self, el: Word)

Adds a new element to the MMR.

pub fn peaks(&self) -> MmrPeaks

Returns the current peaks of the MMR.

pub fn peaks_at(&self, forest: Forest) -> Result<MmrPeaks, MmrError>

Returns the peaks of the MMR at the state specified by forest.

Errors

Returns an error if the specified forest value is not valid for this MMR.

pub fn get_delta( &self, from_forest: Forest, to_forest: Forest, ) -> Result<MmrDelta, MmrError>

Compute the required update to original_forest.

The result is a packed sequence of the authentication elements required to update the trees that have been merged together, followed by the new peaks of the Mmr.

pub fn inner_nodes(&self) -> MmrNodes<'_>

An iterator over inner nodes in the MMR. The order of iteration is unspecified.