miden_crypto::merkle

Module smt

Sparse Merkle Tree (SMT) data structures.

Structs

ForestConfig

The configuration for the forest’s behavior.

ForestInMemoryBackend

The in-memory backend itself.

LargeSmt

A large-scale Sparse Merkle tree mapping 256-bit keys to 256-bit values, backed by pluggable storage. Both keys and values are represented by 4 field elements.

LargeSmtForest

A high-performance forest of sparse merkle trees with pluggable storage backends.

LeafIndex

The index of a leaf, at a depth known at compile-time.

LineageId

An identifier for a lineage of trees.

MemoryStorage

In-memory storage for a Sparse Merkle Tree (SMT), implementing the SmtStorage trait.

MutationSet

Represents a group of prospective mutations to a SparseMerkleTree, created by SparseMerkleTree::compute_mutations(), and that can be applied with SparseMerkleTree::apply_mutations().

PartialSmt

A partial version of an super::Smt.

RocksDbConfig

Configuration for RocksDB storage used by the Sparse Merkle Tree implementation.

RocksDbStorage

A RocksDB-backed persistent storage implementation for a Sparse Merkle Tree (SMT).

SimpleSmt

A sparse Merkle tree with 64-bit keys and 4-element leaf values, without compaction.

SimpleSmtProof

A container for a crate::Word value and its SparseMerklePath opening.

Smt

Sparse Merkle tree mapping 256-bit keys to 256-bit values. Both keys and values are represented by 4 field elements.

SmtForest

An in-memory data collection of sparse Merkle trees (SMTs).

SmtForestUpdateBatch

A batch of operations that can be performed on an arbitrary forest, consisting of operations associated with specified trees in that forest.

SmtProof

A proof which can be used to assert presence (or absence) of key-value pairs in a super::Smt (Sparse Merkle Tree).

SmtUpdateBatch

A batch of operations that can be performed on an arbitrary tree in a forest.

StorageUpdateParts

Owned decomposition of a StorageUpdates batch into its constituent parts.

StorageUpdates

Represents a collection of changes to be applied atomically to an SMT storage backend.

Subtree

Represents a complete 8-depth subtree that is serialized into a single RocksDB entry.

SubtreeLeaf

Helper struct for organizing the data we care about when computing Merkle subtrees.

TreeEntry

An entry in a given tree.

TreeId

An identifier that is capable of uniquely referring to a tree in the forest.

TreeWithRoot

The unique identifier for a given tree, along with the value of its root.

Enums

BackendError

The error type for use within Backends.

ForestOperation

The operations that can be performed on an arbitrary leaf in a tree in a forest.

LargeSmtError

Errors that can occur during LargeSmt operations.

LargeSmtForestError

The type of errors returned by operations on the large SMT forest.

NodeMutation

A change to an inner node of a sparse Merkle tree that hasn’t yet been applied. MutationSet stores this type in relation to a NodeIndex to keep track of what changes need to occur at which node indices.

RootInfo

Information about the role that a queried root plays in the forest.

SmtLeaf

Represents a leaf node in the Sparse Merkle Tree.

SmtLeafError

Errors that can occur when constructing or validating SMT leaves.

SmtProofError

Errors that can occur when validating SMT proofs.

SubtreeError

Constants

FOREST_DEFAULT_MAX_HISTORY_VERSIONS

The default number of historical versions of each tree to keep.

FOREST_MIN_HISTORY_VERSIONS

The minimum number of historical versions per lineage that the forest can store.

MAX_LEAF_ENTRIES

The maximum number of entries allowed in a multiple leaf.

SMT_DEPTH

The depth of the sparse Merkle tree.

SMT_MAX_DEPTH

Maximum supported depth.

SMT_MIN_DEPTH

Minimum supported depth.

Traits

Backend

The backing storage for the SMT forest, providing the necessary high-level methods for performing operations on the full trees that make up the forest, while allowing the forest itself to be storage agnostic.

SmtStorage

Sparse Merkle Tree storage backend.

Functions

build_subtree_for_bench

Type Aliases

VersionId

An identifier for the version of a tree in a given lineage