miden_crypto::merkle::smt

Enum SmtLeaf

pub enum SmtLeaf {
    Empty(LeafIndex<SMT_DEPTH>),
    Single((Word, Word)),
    Multiple(Vec<(Word, Word)>),
}

Represents a leaf node in the Sparse Merkle Tree.

A leaf can be empty, hold a single key-value pair, or multiple key-value pairs.

Variants

Empty(LeafIndex<SMT_DEPTH>)

An empty leaf at the specified index.

Single((Word, Word))

A leaf containing a single key-value pair.

Multiple(Vec<(Word, Word)>)

A leaf containing multiple key-value pairs.

Implementations

impl SmtLeaf

pub fn new( entries: Vec<(Word, Word)>, leaf_index: LeafIndex<SMT_DEPTH>, ) -> Result<Self, SmtLeafError>

Returns a new leaf with the specified entries

Errors

Returns an error if 2 keys in entries map to a different leaf index

Returns an error if 1 or more keys in entries map to a leaf index different from leaf_index

pub fn new_empty(leaf_index: LeafIndex<SMT_DEPTH>) -> Self

Returns a new empty leaf with the specified leaf index

pub fn new_single(key: Word, value: Word) -> Self

Returns a new single leaf with the specified entry. The leaf index is derived from the entryโs key.

pub fn new_multiple(entries: Vec<(Word, Word)>) -> Result<Self, SmtLeafError>

Returns a new multiple leaf with the specified entries. The leaf index is derived from the entriesโ keys.

Errors

Returns an error if 2 keys in entries map to a different leaf index

Returns an error if the number of entries exceeds MAX_LEAF_ENTRIES

pub fn is_empty(&self) -> bool

Returns true if the leaf is empty

pub fn index(&self) -> LeafIndex<SMT_DEPTH>

Returns the leafโs index in the super::Smt

pub fn num_entries(&self) -> usize

Returns the number of entries stored in the leaf

pub fn hash(&self) -> Word

Computes the hash of the leaf

pub fn entries(&self) -> &[(Word, Word)]

Returns a slice with key-value pairs in the leaf.

pub fn to_elements(&self) -> impl Iterator<Item = Felt> + '_

Returns an iterator over the field elements representing this leaf.

pub fn into_elements(self) -> Vec<Felt>

Converts a leaf to a list of field elements.

pub fn into_entries(self) -> Vec<(Word, Word)>

Converts a leaf the key-value pairs in the leaf

pub fn try_from_elements( elements: &[Felt], leaf_index: LeafIndex<SMT_DEPTH>, ) -> Result<SmtLeaf, SmtLeafError>

Converts a list of elements into a leaf