miden_crypto::merkle::smt

Struct ForestPersistentBackend

pub struct ForestPersistentBackend { /* private fields */ }

The persistent backend for the SMT forest, providing durable storage for the latest tree in each lineage in the forest.

Implementations

impl PersistentBackend

This block contains functions for the construction of the persistent backend.

pub fn load(config: Config) -> Result<Self, BackendError>

Constructs an instance of the persistent backend, either opening or creating the data store at the location specified in the config.

Errors

BackendError::CorruptedData if data corruption is encountered when loading the forest from disk.

BackendError::Internal if the backend cannot be started up properly.