miden_crypto::merkle::smt

Struct LargeSmtForest

pub struct LargeSmtForest<B: Backend> { /* private fields */ }

A high-performance forest of sparse merkle trees with pluggable storage backends.

See the module documentation for more information.

Implementations

impl<B: Backend> LargeSmtForest<B>

These functions deal with the creation of new forest instances, and hence rely on the ability to query the backend to do so.

Performance

All the methods in this impl block require access to the underlying Backend instance to return results. This means that their performance will depend heavily on the specific instance with which the forest was constructed.

Where anything more specific can be said about performance, the method documentation will contain more detail.

pub fn new(backend: B) -> Result<Self, LargeSmtForestError>

Constructs a new forest backed by the provided backend using the default Config for the forest’s behavior.

This constructor will treat whatever state is contained within the provided backend as the starting state for the forest. This means that, if you pass a newly-initialized storage, the forest will start in an empty state. Similarly, if you pass a backend that already contains some data (loaded from disk, for example), then the forest will start in that state instead.

Performance

For performance notes on this method, see Self::with_config instead.

Errors

LargeSmtForestError::Other if the forest cannot be started up correctly using the provided backend.

pub fn with_config( backend: B, config: Config, ) -> Result<Self, LargeSmtForestError>

Constructs a new forest backed by the provided backend and configuring behavior using the provided config.

This constructor will treat whatever state is contained within the provided backend as the starting state for the forest. This means that, if you pass a newly-initialized storage, the forest will start in an empty state. Similarly, if you pass a backend that already contains some data (loaded from disk, for example), then the forest will start in that state instead.

Performance

This method is required to load the basic tree metadata from the backend during forest construction. This metadata should be stored separately, and hence this method should take a relatively small amount of time.

Errors

LargeSmtForestError::Fatal if the forest cannot be started up correctly using the provided backend.

impl<B: Backend> LargeSmtForest<B>

These methods provide the ability to perform basic operations on the forest without the need to query the backend.

Performance

All of these methods can be performed fully in-memory, and hence their performance is predictable on a given machine regardless of the choice of Backend instance being used by the forest.

pub fn roots(&self) -> impl Iterator<Item = UniqueRoot>

Returns an iterator that yields all the (uniquely identified) roots that the forest knows about, including those from historical versions.

The iteration order of these roots is unspecified.

pub fn latest_version(&self, lineage: LineageId) -> Option<VersionId>

Gets the latest version of the tree for the provided lineage, if that lineage is in the forest, or returns None otherwise.

pub fn lineage_roots( &self, lineage: LineageId, ) -> Option<impl Iterator<Item = Word>>

Returns an iterator that yields the root values for trees within the specified lineage, or None if the lineage is not known.

The iteration order of the roots is guaranteed to move backward in time as the iterator advances, with earlier items being roots from versions closer to the present. The current root of the lineage will thus always be the first item yielded by the iterator.

pub fn latest_root(&self, lineage: LineageId) -> Option<Word>

Gets the value root of the newest tree in the provided lineage, if that lineage is in the forest, or returns None otherwise.

pub fn tree_count(&self) -> usize

Returns the number of trees in the forest that have unique identity.

This is not the number of unique tree lineages in the forest, as it includes all historical trees as well. For that, see Self::lineage_count.

pub fn lineage_count(&self) -> usize

Returns the number of unique tree lineages in the forest.

This is not the number of unique trees in the forest, as it does not include all versions in each lineage. For that, see Self::tree_count.

pub fn root_info(&self, root: TreeId) -> RootInfo

Returns data describing what information the forest knows about the provided root.

pub fn truncate(&mut self, version: VersionId)

Removes all tree versions in the forest that are older than the provided version, but always retains the latest tree in each lineage.

impl<B: Backend> LargeSmtForest<B>

These methods pertain to non-mutating queries about the data stored in the forest. They differ from the simple queries in the previous block by requiring access to the backend to function.

Performance

All the methods in this impl block require access to the underlying Backend instance to return results. This means that their performance will depend heavily on the specific instance with which the forest was constructed.

Where anything more specific can be said about performance, the method documentation will contain more detail.

pub fn open( &self, tree: TreeId, key: Word, ) -> Result<SmtProof, LargeSmtForestError>

Returns an opening for the specified key in the specified tree, regardless of whether the tree has a value associated with key or not.

Errors

LargeSmtForestError::Fatal if the backend fails to operate properly during the query.

LargeSmtForestError::UnknownLineage If the provided tree specifies a lineage that is not one known by the forest.

LargeSmtForestError::UnknownTree if the provided tree refers to a tree that is not a member of the forest.

LargeSmtForestError::Merkle if there is insufficient data in the specified tree to provide an opening for key.

pub fn get( &self, tree: TreeId, key: Word, ) -> Result<Option<Word>, LargeSmtForestError>

Returns the value associated with the provided key in the specified tree, or None if there is no non-default value corresponding to the provided key in that tree.

Errors

LargeSmtForestError::Fatal if the backend fails to operate properly during the query.

LargeSmtForestError::UnknownLineage if the provided tree specifies a lineage that is not one known by the forest.

LargeSmtForestError::UnknownTree if the provided tree refers to a tree that is not a member of the forest.

pub fn entry_count(&self, tree: TreeId) -> Result<usize, LargeSmtForestError>

Returns the number of populated entries in the specified tree.

Performance

Due to the way that tree data is stored, this method exhibits a split performance profile.

If querying for a tree that is the latest in its lineage, the time to return a result should be constant.

If querying for a tree that is a historical version, the time to return a result will be linear in the number of entries in the tree. This is because an overlaid iterator has to be created to yield the correct entries for the historical version, and then queried for its length.

Errors

LargeSmtForestError::Fatal if the backend fails to operate properly during the query.

LargeSmtForestError::UnknownLineage if the provided tree specifies a lineage that is not one known by the forest.

LargeSmtForestError::UnknownTree if the provided tree refers to a tree that is not a member of the forest.

pub fn entries( &self, tree: TreeId, ) -> Result<impl Iterator<Item = TreeEntry>, LargeSmtForestError>

Returns an iterator that yields the entries in the specified tree.

Performance

The performance of the iterator depends both on the choice of backend and the type of tree that is queried for. We cannot give exact performance figures, but in general querying over the current tree in a lineage will be faster than querying over a historical tree in a lineage.

Errors

LargeSmtForestError::Fatal if the backend fails to operate properly during the query.

LargeSmtForestError::UnknownLineage if the provided tree specifies a lineage that is not one known by the forest.

LargeSmtForestError::UnknownTree if the provided tree refers to a tree that is not a member of the forest.

impl<B: Backend> LargeSmtForest<B>

These methods pertain to modifications that can be made to a single tree in the forest. They exploit parallelism within the single target tree wherever possible.

Performance

All the methods in this impl block require access to the underlying Backend instance to return results. This means that their performance will depend heavily on the specific instance with which the forest was constructed.

Where anything more specific can be said about performance, the method documentation will contain more detail.

pub fn add_lineage( &mut self, lineage: LineageId, new_version: VersionId, updates: SmtUpdateBatch, ) -> Result<TreeWithRoot, LargeSmtForestError>

Adds a new lineage to the tree, creating an empty tree and modifying it as specified by updates, with the result taking the provided new_version.

If the provided updates batch is empty, then the empty tree will be added as the first version in the lineage.

Errors

LargeSmtForestError::DuplicateLineage if the provided lineage is the same as an already-known lineage.

LargeSmtForestError::Fatal if the backend fails while being accessed.

BackendError::Merkle if the provided updates cannot be applied to the empty tree.

pub fn update_tree( &mut self, lineage: LineageId, new_version: VersionId, updates: SmtUpdateBatch, ) -> Result<TreeWithRoot, LargeSmtForestError>

Performs the provided updates on the latest tree in the specified lineage, adding a single new root to the forest (corresponding to new_version) for the entire batch, and returning the data for the new root of the tree.

If applying the provided operations results in no changes to the tree, then the root data will be returned unchanged and no new tree will be allocated. It will retain its original version, and not be returned with new_version.

Errors

LargeSmtForestError::BadVersion if the new_version is older than the latest version for the provided lineage.

LargeSmtForestError::Fatal if the backend fails while being accessed.

LargeSmtForestError::UnknownLineage if the provided tree specifies a lineage that is not one known by the forest.

impl<B: Backend> LargeSmtForest<B>

These methods pertain to modifications that can be made to multiple trees in the forest at once. They exploit parallelism both between trees and within trees wherever possible.

Performance

All the methods in this impl block require access to the underlying Backend instance to return results. This means that their performance will depend heavily on the specific instance with which the forest was constructed.

Where anything more specific can be said about performance, the method documentation will contain more detail.

pub fn update_forest( &mut self, new_version: VersionId, updates: SmtForestUpdateBatch, ) -> Result<Vec<TreeWithRoot>, LargeSmtForestError>

Performs the provided updates on the forest, adding at most one new root with version new_version to the forest for each target root in updates and returning a mapping from old root to the new root data.

If applying the associated batch to any given lineage in the forest results in no changes to that tree, the initial root for that lineage will be returned and no new tree will be allocated.

Errors

LargeSmtForestError::UnknownLineage If any lineage in the batch of modifications is one that is not known by the forest.

LargeSmtForestError::Fatal if any error occurs to leave the forest in an inconsistent state.

LargeSmtForestError::BadVersion if the new_version is older than the latest version for the provided lineage.