miden_crypto::merkle

Struct SparseMerklePath

pub struct SparseMerklePath { /* private fields */ }

A different representation of MerklePath designed for memory efficiency for Merkle paths with empty nodes.

Empty nodes in the path are stored only as their position, represented with a bitmask. A maximum of 64 nodes (SMT_MAX_DEPTH) can be stored (empty and non-empty). The more nodes in a path are empty, the less memory this struct will use. This type calculates empty nodes on-demand when iterated through, converted to a MerklePath, or an empty node is retrieved with SparseMerklePath::at_depth(), which will incur overhead.

NOTE: This type assumes that Merkle paths always span from the root of the tree to a leaf. Partial paths are not supported.

Implementations

impl SparseMerklePath

pub fn from_parts( empty_nodes_mask: u64, nodes: Vec<Word>, ) -> Result<Self, MerkleError>

Constructs a new sparse Merkle path from a bitmask of empty nodes and a vector of non-empty nodes.

The empty_nodes_mask is a bitmask where each set bit indicates that the node at that depth is empty. The least significant bit (bit 0) describes depth 1 node (root’s children). The bit index + 1 is equal to node’s depth. The nodes vector must contain the non-empty nodes in depth order.

Errors

MerkleError::InvalidPathLength if the provided nodes vector is shorter than the minimum length required by the empty_nodes_mask.

MerkleError::DepthTooBig if the total depth of the path (calculated from the empty_nodes_mask and nodes) is greater than SMT_MAX_DEPTH.

pub fn from_sized_iter<I>(iterator: I) -> Result<Self, MerkleError>
where I: IntoIterator<IntoIter: ExactSizeIterator, Item = Word>,

Constructs a sparse Merkle path from an iterator over Merkle nodes that also knows its exact size (such as iterators created with Vec::into_iter). The iterator must be in order of deepest to shallowest.

Knowing the size is necessary to calculate the depth of the tree, which is needed to detect which nodes are empty nodes.

Errors

Returns MerkleError::DepthTooBig if tree_depth is greater than SMT_MAX_DEPTH.

pub fn depth(&self) -> u8

Returns the total depth of this path, i.e., the number of nodes this path represents.

pub fn at_depth(&self, node_depth: NonZero<u8>) -> Result<Word, MerkleError>

Get a specific node in this path at a given depth.

The depth parameter is defined in terms of self.depth(). Merkle paths conventionally do not include the root, so the shallowest depth is 1, and the deepest depth is self.depth().

Errors

Returns MerkleError::DepthTooBig if node_depth is greater than the total depth of this path.

pub fn into_parts(self) -> (u64, Vec<Word>)

Deconstructs this path into its component parts.

Returns a tuple containing:

a bitmask where each set bit indicates that the node at that depth is empty. The least significant bit (bit 0) describes depth 1 node (root’s children).

a vector of non-empty nodes in depth order.

pub fn iter(&self) -> impl ExactSizeIterator<Item = Word>

Constructs a borrowing iterator over the nodes in this path. Starts from the leaf and iterates toward the root (excluding the root).

pub fn compute_root( &self, index: u64, node_to_prove: Word, ) -> Result<Word, MerkleError>

Computes the Merkle root for this opening.

pub fn verify( &self, index: u64, node: Word, expected_root: &Word, ) -> Result<(), MerkleError>

Verifies the Merkle opening proof towards the provided root.

Errors

Returns an error if:

provided node index is invalid.

root calculated during the verification differs from the provided one.

pub fn authenticated_nodes( &self, index: u64, node_to_prove: Word, ) -> Result<InnerNodeIterator<'_>, MerkleError>

Given the node this path opens to, return an iterator of all the nodes that are known via this path.

Each item in the iterator is an InnerNodeInfo, containing the hash of a node as .value, and its two children as .left and .right. The very first item in that iterator will be the parent of node_to_prove as stored in this SparseMerklePath.

From there, the iterator will continue to yield every further parent and both of its children, up to and including the root node.

If node_to_prove is not the node this path is an opening to, or index is not the correct index for that node, the returned nodes will be meaningless.

Errors

Returns an error if the specified index is not valid for this path.