miden_crypto::rand

Struct RpxRandomCoin

pub struct RpxRandomCoin { /* private fields */ }

A simplified version of the SPONGE_PRG reseedable pseudo-random number generator algorithm described in https://eprint.iacr.org/2011/499.pdf.

The simplification is related to the following facts:

A call to the reseed method implies one and only one call to the permutation function. This is possible because in our case we never reseed with more than 4 field elements.

As a result of the previous point, we donโt make use of an input buffer to accumulate seed material.

Implementations

impl RpxRandomCoin

pub fn new(seed: Word) -> Self

Returns a new RpxRandomCoin initialize with the specified seed.

pub fn from_parts(state: [Felt; 12], current: usize) -> Self

Returns an RpxRandomCoin instantiated from the provided components.

Panics

Panics if current is smaller than 4 or greater than or equal to 12.

pub fn into_parts(self) -> ([Felt; 12], usize)

Returns components of this random coin.

pub fn fill_bytes(&mut self, dest: &mut [u8])

Fills dest with random data.

pub fn draw_basefield(&mut self) -> Felt

pub fn draw(&mut self) -> Felt

Draws a random field element.

This is an alias for Self::draw_basefield.

pub fn draw_ext_field<E: ExtensionField<Felt>>(&mut self) -> E

pub fn reseed(&mut self, data: Word)

pub fn check_leading_zeros(&self, value: u64) -> u32

pub fn draw_integers( &mut self, num_values: usize, domain_size: usize, nonce: u64, ) -> Vec<usize>