miden_crypto::rand

Trait FeltRng

pub trait FeltRng: RngCore {
    // Required methods
    fn draw_element(&mut self) -> Felt;
    fn draw_word(&mut self) -> Word;
}

Pseudo-random element generator.

An instance can be used to draw, uniformly at random, base field elements as well as Words.

Required Methods

fn draw_element(&mut self) -> Felt

Draw, uniformly at random, a base field element.

fn draw_word(&mut self) -> Word

Draw, uniformly at random, a Word.