miden_crypto::stark::air

Struct FilteredMidenAirBuilder

pub struct FilteredMidenAirBuilder<'a, AB>
where
    AB: MidenAirBuilder,
{
    pub inner: &'a mut AB,
    pub condition: <AB as MidenAirBuilder>::Expr,
}

A wrapper around a MidenAirBuilder that enforces constraints only when a specified condition is met.

This struct allows selectively applying constraints to certain rows or under certain conditions in the AIR, without modifying the underlying logic. All constraints asserted through this filtered builder will be multiplied by the given condition, effectively disabling them when condition evaluates to zero.

Fields
inner: &'a mut AB

Reference to the underlying inner MidenAirBuilder where constraints are ultimately recorded.

condition: <AB as MidenAirBuilder>::Expr

Condition expression that controls when the constraints are enforced.

If condition evaluates to zero, constraints asserted through this builder have no effect.

Implementations

impl<AB> FilteredMidenAirBuilder<'_, AB>
where AB: MidenAirBuilder,

pub fn condition(&self) -> <AB as MidenAirBuilder>::Expr