miden_crypto::stark::air

Trait BaseAirWithAuxTrace

pub trait BaseAirWithAuxTrace<F, EF>: BaseAir<F> {
    // Provided methods
    fn num_randomness(&self) -> usize { ... }
    fn aux_width(&self) -> usize { ... }
    fn bus_types(&self) -> &[BusType] { ... }
    fn build_aux_trace(
        &self,
        _main: &DenseMatrix<F>,
        _challenges: &[EF],
    ) -> Option<DenseMatrix<F>> { ... }
    fn with_aux_builder<Builder>(&mut self, _builder: Builder)
       where Builder: Fn(&DenseMatrix<F>, &[EF]) -> DenseMatrix<F> + Send + Sync + 'static { ... }
}

An extension of BaseAir that includes support for auxiliary traces.

This trait is defined in p3-miden-air (not upstream p3-air) because auxiliary trace support is Miden-specific functionality.

Provided Methods

fn num_randomness(&self) -> usize

Number of challenges (extension fields) that is required to compute the aux trace

fn aux_width(&self) -> usize

Number of columns (in extension field) that is required for aux trace

fn bus_types(&self) -> &[BusType]

Types of buses

fn build_aux_trace( &self, _main: &DenseMatrix<F>, _challenges: &[EF], ) -> Option<DenseMatrix<F>>

Build an aux trace (EF-based) given the main trace and EF challenges. Return None to indicate no aux or to fall back to legacy behavior.

fn with_aux_builder<Builder>(&mut self, _builder: Builder)
where Builder: Fn(&DenseMatrix<F>, &[EF]) -> DenseMatrix<F> + Send + Sync + 'static,

Load an aux builder.

An aux builder takes in a main matrix and a randomness, and generate a aux matrix.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.