miden_crypto::stark::air

Trait MidenAirBuilder

pub trait MidenAirBuilder: Sized {
    type F: Field + Sync;
    type Expr: Algebra<Self::F> + Algebra<Self::Var>;
    type Var: Into<Self::Expr> + Clone + Send + Sync + Add<Self::F, Output = Self::Expr, Output = Self::Expr, Output = Self::Expr> + Add + Add<Self::Expr> + Sub<Self::F, Output = Self::Expr, Output = Self::Expr, Output = Self::Expr> + Sub + Sub<Self::Expr> + Mul<Self::F, Output = Self::Expr, Output = Self::Expr, Output = Self::Expr> + Mul + Mul<Self::Expr>;
    type M: Matrix<Self::Var>;
    type PublicVar: Into<Self::Expr> + Copy;
    type PeriodicVal: Into<Self::ExprEF> + Copy;
    type EF: ExtensionField<Self::F>;
    type ExprEF: Algebra<Self::Expr> + Algebra<Self::EF>;
    type VarEF: Into<Self::ExprEF> + Copy + Send + Sync;
    type MP: Matrix<Self::VarEF>;
    type RandomVar: Into<Self::ExprEF> + Copy;

    // Required methods
    fn main(&self) -> Self::M;
    fn is_first_row(&self) -> Self::Expr;
    fn is_last_row(&self) -> Self::Expr;
    fn is_transition_window(&self, size: usize) -> Self::Expr;
    fn assert_zero<I>(&mut self, x: I)
       where I: Into<Self::Expr>;
    fn public_values(&self) -> &[Self::PublicVar];
    fn periodic_evals(&self) -> &[Self::PeriodicVal];
    fn preprocessed(&self) -> Self::M;
    fn assert_zero_ext<I>(&mut self, x: I)
       where I: Into<Self::ExprEF>;
    fn permutation(&self) -> Self::MP;
    fn permutation_randomness(&self) -> &[Self::RandomVar];
    fn aux_bus_boundary_values(&self) -> &[Self::VarEF];

    // Provided methods
    fn is_transition(&self) -> Self::Expr { ... }
    fn when<I>(&mut self, condition: I) -> FilteredMidenAirBuilder<'_, Self>
       where I: Into<Self::Expr> { ... }
    fn when_ne<I1, I2>(
        &mut self,
        x: I1,
        y: I2,
    ) -> FilteredMidenAirBuilder<'_, Self>
       where I1: Into<Self::Expr>,
             I2: Into<Self::Expr> { ... }
    fn when_first_row(&mut self) -> FilteredMidenAirBuilder<'_, Self> { ... }
    fn when_last_row(&mut self) -> FilteredMidenAirBuilder<'_, Self> { ... }
    fn when_transition(&mut self) -> FilteredMidenAirBuilder<'_, Self> { ... }
    fn when_transition_window(
        &mut self,
        size: usize,
    ) -> FilteredMidenAirBuilder<'_, Self> { ... }
    fn assert_zeros<const N: usize, I>(&mut self, array: [I; N])
       where I: Into<Self::Expr> { ... }
    fn assert_bools<const N: usize, I>(&mut self, array: [I; N])
       where I: Into<Self::Expr> { ... }
    fn assert_one<I>(&mut self, x: I)
       where I: Into<Self::Expr> { ... }
    fn assert_eq<I1, I2>(&mut self, x: I1, y: I2)
       where I1: Into<Self::Expr>,
             I2: Into<Self::Expr> { ... }
    fn assert_bool<I>(&mut self, x: I)
       where I: Into<Self::Expr> { ... }
    fn assert_eq_ext<I1, I2>(&mut self, x: I1, y: I2)
       where I1: Into<Self::ExprEF>,
             I2: Into<Self::ExprEF> { ... }
    fn assert_one_ext<I>(&mut self, x: I)
       where I: Into<Self::ExprEF> { ... }

}

Super trait for all AIR builders in the Miden VM ecosystem.

This trait contains all methods from AirBuilder, AirBuilderWithPublicValues, PairBuilder, ExtensionBuilder, and PermutationAirBuilder. Implementers only need to implement this single trait.

To use your builder with the STARK prover/verifier, you’ll also need to implement the p3-air traits using the impl_p3_air_builder_traits! macro.

Type Parameters

All type parameters are associated types that must be specified by implementers.

Required Methods

Core methods that must be implemented:

main() - Access to main trace

is_first_row() - First row indicator

is_last_row() - Last row indicator

is_transition_window() - Transition window indicator

assert_zero() - Core constraint assertion

public_values() - Access to public values

preprocessed() - Access to preprocessed columns

assert_zero_ext() - Extension field constraint assertion

permutation() - Access to permutation trace

permutation_randomness() - Access to randomness

Optional Methods (with default implementations)

All other methods have default implementations.

Required Associated Types

type F: Field + Sync

Underlying field type.

type Expr: Algebra<Self::F> + Algebra<Self::Var>

Serves as the output type for an AIR constraint evaluation.

type Var: Into<Self::Expr> + Clone + Send + Sync + Add<Self::F, Output = Self::Expr, Output = Self::Expr, Output = Self::Expr> + Add + Add<Self::Expr> + Sub<Self::F, Output = Self::Expr, Output = Self::Expr, Output = Self::Expr> + Sub + Sub<Self::Expr> + Mul<Self::F, Output = Self::Expr, Output = Self::Expr, Output = Self::Expr> + Mul + Mul<Self::Expr>

The type of the variable appearing in the trace matrix.

type M: Matrix<Self::Var>

Matrix type holding variables.

type PublicVar: Into<Self::Expr> + Copy

Type representing a public variable.

type PeriodicVal: Into<Self::ExprEF> + Copy

Type representing a periodic value.

type EF: ExtensionField<Self::F>

Extension field type.

type ExprEF: Algebra<Self::Expr> + Algebra<Self::EF>

Expression type over extension field elements.

type VarEF: Into<Self::ExprEF> + Copy + Send + Sync

Variable type over extension field elements.

type MP: Matrix<Self::VarEF>

Matrix type over extension-field variables representing permutation registers.

type RandomVar: Into<Self::ExprEF> + Copy

Randomness variable type used in permutation commitments.

Required Methods

fn main(&self) -> Self::M

Return the matrix representing the main (primary) trace registers.

fn is_first_row(&self) -> Self::Expr

Expression evaluating to 1 on the first row, 0 elsewhere.

fn is_last_row(&self) -> Self::Expr

Expression evaluating to 1 on the last row, 0 elsewhere.

fn is_transition_window(&self, size: usize) -> Self::Expr

Expression evaluating to 1 on rows except the last size - 1 rows, 0 otherwise.

fn assert_zero<I>(&mut self, x: I)
where I: Into<Self::Expr>,

Assert that the given element is zero.

Where possible, batching multiple assert_zero calls into a single assert_zeros call will improve performance.

fn public_values(&self) -> &[Self::PublicVar]

Return the list of public variables.

fn periodic_evals(&self) -> &[Self::PeriodicVal]

Return the list of periodic values.

fn preprocessed(&self) -> Self::M

Return a matrix of preprocessed registers.

fn assert_zero_ext<I>(&mut self, x: I)
where I: Into<Self::ExprEF>,

Assert that an extension field expression is zero.

fn permutation(&self) -> Self::MP

Return the matrix representing permutation registers over EF variables.

fn permutation_randomness(&self) -> &[Self::RandomVar]

Return the list of randomness values for permutation argument.

fn aux_bus_boundary_values(&self) -> &[Self::VarEF]

Aux bus boundary values: EF finals, one per aux/bus column, carried in the proof.

Provided Methods

fn is_transition(&self) -> Self::Expr

Expression evaluating to 1 on all transition rows (not last row), 0 on last row.

fn when<I>(&mut self, condition: I) -> FilteredMidenAirBuilder<'_, Self>
where I: Into<Self::Expr>,

Returns a sub-builder whose constraints are enforced only when condition is nonzero.

fn when_ne<I1, I2>(&mut self, x: I1, y: I2) -> FilteredMidenAirBuilder<'_, Self>
where I1: Into<Self::Expr>, I2: Into<Self::Expr>,

Returns a sub-builder whose constraints are enforced only when x != y.

fn when_first_row(&mut self) -> FilteredMidenAirBuilder<'_, Self>

Returns a sub-builder whose constraints are enforced only on the first row.

fn when_last_row(&mut self) -> FilteredMidenAirBuilder<'_, Self>

Returns a sub-builder whose constraints are enforced only on the last row.

fn when_transition(&mut self) -> FilteredMidenAirBuilder<'_, Self>

Returns a sub-builder whose constraints are enforced on all rows except the last.

fn when_transition_window( &mut self, size: usize, ) -> FilteredMidenAirBuilder<'_, Self>

Returns a sub-builder whose constraints are enforced on all rows except the last size - 1.

fn assert_zeros<const N: usize, I>(&mut self, array: [I; N])
where I: Into<Self::Expr>,

Assert that every element of a given array is 0.

This should be preferred over calling assert_zero multiple times.

fn assert_bools<const N: usize, I>(&mut self, array: [I; N])
where I: Into<Self::Expr>,

Assert that a given array consists of only boolean values.

fn assert_one<I>(&mut self, x: I)
where I: Into<Self::Expr>,

Assert that x element is equal to 1.

fn assert_eq<I1, I2>(&mut self, x: I1, y: I2)
where I1: Into<Self::Expr>, I2: Into<Self::Expr>,

Assert that the given elements are equal.

fn assert_bool<I>(&mut self, x: I)
where I: Into<Self::Expr>,

Assert that x is a boolean, i.e. either 0 or 1.

Where possible, batching multiple assert_bool calls into a single assert_bools call will improve performance.

fn assert_eq_ext<I1, I2>(&mut self, x: I1, y: I2)
where I1: Into<Self::ExprEF>, I2: Into<Self::ExprEF>,

Assert that two extension field expressions are equal.

fn assert_one_ext<I>(&mut self, x: I)
where I: Into<Self::ExprEF>,

Assert that an extension field expression is equal to one.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.