miden_crypto::stark::commit

Struct MerkleTreeMmcs

pub struct MerkleTreeMmcs<P, PW, H, C, const DIGEST_ELEMS: usize> { /* private fields */ }

A Merkle Tree-based commitment scheme for multiple matrices of potentially differing heights.

MerkleTreeMmcs generalizes a classical Merkle Tree to support committing to a list of matrices by arranging their rows into a unified binary tree. The tallest matrix defines the maximum height, and smaller matrices are integrated at appropriate depths.

Type Parameters:

P: Packed leaf value (e.g. a field element or vector of elements)

PW: Packed digest element (used in the hash and compression output)

H: Cryptographic hash function (leaf hash)

C: Pseudo-compression function (internal node compression)

DIGEST_ELEMS: Number of elements in a single digest

Implementations

impl<P, PW, H, C, const DIGEST_ELEMS: usize> MerkleTreeMmcs<P, PW, H, C, DIGEST_ELEMS>

pub const fn new( hash: H, compress: C, ) -> MerkleTreeMmcs<P, PW, H, C, DIGEST_ELEMS>

Create a new MerkleTreeMmcs with the given hash and compression functions.