miden_crypto::stark::pcs

Struct TwoAdicFriPcs

pub struct TwoAdicFriPcs<Val, Dft, InputMmcs, FriMmcs> { /* private fields */ }

A polynomial commitment scheme using FRI to generate opening proofs.

We commit to a polynomial f via its evaluation vectors over a coset gH where |H| >= 2 * deg(f). A value f(z) is opened by using a FRI proof to show that the evaluations of (f(x) - f(z))/(x - z) over gH are low degree.

Implementations

impl<Val, Dft, InputMmcs, FriMmcs> TwoAdicFriPcs<Val, Dft, InputMmcs, FriMmcs>

pub const fn new( dft: Dft, mmcs: InputMmcs, fri: FriParameters<FriMmcs>, ) -> TwoAdicFriPcs<Val, Dft, InputMmcs, FriMmcs>