miden_crypto::stark

Struct VerifierConstraintFolder

pub struct VerifierConstraintFolder<'a, SC>
where
    SC: StarkGenericConfig,
{
    pub main: VerticalPair<DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>, DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>>,
    pub aux: VerticalPair<DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>, DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>>,
    pub randomness: &'a [<SC as StarkGenericConfig>::Challenge],
    pub aux_bus_boundary_values: &'a [<SC as StarkGenericConfig>::Challenge],
    pub preprocessed: Option<VerticalPair<DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>, DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>>>,
    pub public_values: &'a [<<<SC as StarkGenericConfig>::Pcs as Pcs<<SC as StarkGenericConfig>::Challenge, <SC as StarkGenericConfig>::Challenger>>::Domain as PolynomialSpace>::Val],
    pub periodic_values: &'a [<SC as StarkGenericConfig>::Challenge],
    pub is_first_row: <SC as StarkGenericConfig>::Challenge,
    pub is_last_row: <SC as StarkGenericConfig>::Challenge,
    pub is_transition: <SC as StarkGenericConfig>::Challenge,
    pub alpha: <SC as StarkGenericConfig>::Challenge,
    pub accumulator: <SC as StarkGenericConfig>::Challenge,
}

Handles constraint verification for the verifier in a STARK system.

Similar to ProverConstraintFolder but operates on committed values rather than the full trace, using a more efficient accumulation method for verification.

Fields
main: VerticalPair<DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>, DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>>

Pair of consecutive rows from the committed polynomial evaluations

aux: VerticalPair<DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>, DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>>

Pair of consecutive rows from the committed polynomial evaluations (may have zero width)

randomness: &'a [<SC as StarkGenericConfig>::Challenge]

The randomness used to compute the aux tract; can be zero width.

aux_bus_boundary_values: &'a [<SC as StarkGenericConfig>::Challenge]

Aux trace bus boundary values; can be zero width.

preprocessed: Option<VerticalPair<DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>, DenseMatrix<<SC as StarkGenericConfig>::Challenge, &'a [<SC as StarkGenericConfig>::Challenge]>>>

The preprocessed columns (if any)

public_values: &'a [<<<SC as StarkGenericConfig>::Pcs as Pcs<<SC as StarkGenericConfig>::Challenge, <SC as StarkGenericConfig>::Challenger>>::Domain as PolynomialSpace>::Val]

Public values that are inputs to the computation

periodic_values: &'a [<SC as StarkGenericConfig>::Challenge]

Periodic column values (precomputed for the current row)

is_first_row: <SC as StarkGenericConfig>::Challenge

Evaluations of the Selector polynomial for the first row of the trace

is_last_row: <SC as StarkGenericConfig>::Challenge

Evaluations of the Selector polynomial for the last row of the trace

is_transition: <SC as StarkGenericConfig>::Challenge

Evaluations of the Selector polynomial for rows where transition constraints should be applied

alpha: <SC as StarkGenericConfig>::Challenge

Single challenge value used for constraint combination

accumulator: <SC as StarkGenericConfig>::Challenge

Running accumulator for all constraints