miden_crypto::stark::symmetric

Trait CryptographicPermutation

pub trait CryptographicPermutation<T>: Permutation<T>
where
    T: Clone,
{ }

A permutation thought to be cryptographically secure, in the sense that it is thought to be difficult to distinguish (in a nontrivial way) from a random permutation.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.