miden_crypto

Trait SequentialCommit

pub trait SequentialCommit {
    type Commitment: From<Word>;

    // Required method
    fn to_elements(&self) -> Vec<Felt>;

    // Provided method
    fn to_commitment(&self) -> Self::Commitment { ... }
}

Defines how to compute a commitment to an object represented as a sequence of field elements.

Required Associated Types

type Commitment: From<Word>

A type of the commitment which must be derivable from Word.

Required Methods

fn to_elements(&self) -> Vec<Felt>

Returns a representation of the object as a sequence of fields elements.

Provided Methods

fn to_commitment(&self) -> Self::Commitment

Computes the commitment to the object.

The default implementation of this function uses Poseidon2 hash function to hash the sequence of elements returned from Self::to_elements().