miden_crypto::word

Struct Word

pub struct Word(/* private fields */);

A unit of data consisting of 4 field elements.

Implementations

impl Word

pub const SERIALIZED_SIZE: usize = WORD_SIZE_BYTES

The serialized size of the word in bytes.

pub const fn new(value: [Felt; 4]) -> Self

Creates a new Word from the given field elements.

pub const fn parse(hex: &str) -> Result<Self, &'static str>

Parses a hex string into a new Word.

The input must contain valid hex prefixed with 0x. The input after the prefix must contain between 0 and 64 characters (inclusive).

The input is interpreted to have little-endian byte ordering. Nibbles are interpreted to have big-endian ordering so that “0x10” represents Felt::new(16), not Felt::new(1).

This function is usually used via the word! macro.

use miden_crypto::{Felt, Word, word};
let word = word!("0x1000000000000000200000000000000030000000000000004000000000000000");
assert_eq!(word, Word::new([Felt::new(16), Felt::new(32), Felt::new(48), Felt::new(64)]));

pub const fn empty() -> Self

Returns a new Word consisting of four ZERO elements.

pub fn is_empty(&self) -> bool

Returns true if the word consists of four ZERO elements.

pub fn as_elements(&self) -> &[Felt]

Returns the word as a slice of field elements.

pub fn as_bytes(&self) -> [u8; 32]

Returns the word as a byte array.

pub fn words_as_elements(words: &[Self]) -> &[Felt]

Returns all elements of multiple words as a slice.

pub fn to_hex(&self) -> String

Returns hexadecimal representation of this word prefixed with 0x.

pub fn to_vec(&self) -> Vec<Felt>

Returns internal elements of this word as a vector.