miden_crypto::merkle::smt

Trait Backend

pub trait Backend
where
    Self: Debug,
{
    // Required methods
    fn open(
        &self,
        lineage: LineageId,
        key: Word,
    ) -> Result<SmtProof, BackendError>;
    fn get(
        &self,
        lineage: LineageId,
        key: Word,
    ) -> Result<Option<Word>, BackendError>;
    fn version(&self, lineage: LineageId) -> Result<VersionId, BackendError>;
    fn lineages(&self) -> Result<impl Iterator<Item = LineageId>, BackendError>;
    fn trees(&self) -> Result<impl Iterator<Item = TreeWithRoot>, BackendError>;
    fn entry_count(&self, lineage: LineageId) -> Result<usize, BackendError>;
    fn entries(
        &self,
        lineage: LineageId,
    ) -> Result<impl Iterator<Item = TreeEntry>, BackendError>;
    fn add_lineage(
        &mut self,
        lineage: LineageId,
        version: VersionId,
        updates: SmtUpdateBatch,
    ) -> Result<TreeWithRoot, BackendError>;
    fn update_tree(
        &mut self,
        lineage: LineageId,
        new_version: VersionId,
        updates: SmtUpdateBatch,
    ) -> Result<MutationSet<SMT_DEPTH, Word, Word>, BackendError>;
    fn update_forest(
        &mut self,
        new_version: VersionId,
        updates: SmtForestUpdateBatch,
    ) -> Result<Vec<(LineageId, MutationSet<SMT_DEPTH, Word, Word>)>, BackendError>;
}

The backing storage for the SMT forest, providing the necessary high-level methods for performing operations on the full trees that make up the forest, while allowing the forest itself to be storage agnostic.

Backend Data Storage

Having a generic Backend provides no guarantees to the user about how it stores data and what patterns are used for data access under the hood. It is, however, guaranteed to store only the data necessary to describe the latest state of each tree in the forest.

Error Handling

We separate errors in backend implementations into two semantic categories:

User-Derived Errors: These are errors that arise downstream of data provided by the user. These errors must be signaled by returning an Err variant with an appropriate error.

Internal Errors: These are errors that are not derived from data provided by the user. Signaling such an error is up to the implementation, but can be done through both panicking and returning the BackendError::Internal variant as appropriate. These may leave the backend in an inconsistent state as they are designed to effect program termination or perform it directly.

The only reason that BackendError::Internal exists is to allow certain failures to result in termination at the level of the forest instead of the backend as this can sometimes lead to cleaner logic. If this is not appropriate, a panic is a better option.

Expected Behavior

Certain methods on this trait (e.g. Backend::update_tree) provide behaviors expected for that method. These combine with the following trait-level behavior requirements to become part of the contract of the method, but a portion that cannot be encoded in the type system. Any failure to conform to these expected behaviors is considered a bug in the implementation of the backend, and must be rectified.

The following behavior is expected of all methods in implementations of this trait:

For any failure derived from user input (see User-Derived Errors above), the data and the backend must be left in a consistent state when the error is returned to the caller.

Failures derived from user input (see User-Derived Errors above) must be signaled to the caller by returning a variant of BackendError that is not BackendError::Internal. Methods may place additional constraints on which errors are used to signal certain failures.

Required Methods

fn open(&self, lineage: LineageId, key: Word) -> Result<SmtProof, BackendError>

Returns an opening for the specified key in the SMT with the specified lineage.

It is the responsibility of the forest to ensure lineage existence before querying the backend. The backend must return an error if the lineage does not exist.

fn get( &self, lineage: LineageId, key: Word, ) -> Result<Option<Word>, BackendError>

Returns the value associated with the provided key in the SMT with the specified lineage, or None if no such value exists.

It is the responsibility of the forest to ensure lineage existence before querying the backend. The backend must return an error if the lineage does not exist.

fn version(&self, lineage: LineageId) -> Result<VersionId, BackendError>

Returns the version of the tree with the specified lineage.

It is the responsibility of the forest to ensure lineage existence before querying the backend. The backend must return an error if the lineage does not exist.

fn lineages(&self) -> Result<impl Iterator<Item = LineageId>, BackendError>

Returns an iterator over all the lineages that the backend knows about.

The iteration order is unspecified.

fn trees(&self) -> Result<impl Iterator<Item = TreeWithRoot>, BackendError>

Returns an iterator over all the trees (and their corresponding roots) that the backend knows about.

The iteration order is unspecified.

fn entry_count(&self, lineage: LineageId) -> Result<usize, BackendError>

Returns the total number of (key-value) entries in the specified lineage.

It is the responsibility of the forest to ensure lineage existence before querying the backend. The backend must return an error if the lineage does not exist.

fn entries( &self, lineage: LineageId, ) -> Result<impl Iterator<Item = TreeEntry>, BackendError>

Returns an iterator that yields the populated (key-value) entries for the specified lineage.

It is the responsibility of the forest to ensure lineage existence before querying the backend. The backend must return an error if the lineage does not exist.

This iterator must yield entries in an order such that they are sorted by their leaf index, and entries that share a leaf index are sorted by key. It must not include key-value pairs where the value is the empty word.

fn add_lineage( &mut self, lineage: LineageId, version: VersionId, updates: SmtUpdateBatch, ) -> Result<TreeWithRoot, BackendError>

Adds a new lineage to the forest with the provided version and sets the associated SMT to have the value created by applying updates to the empty tree, returning the new root of that tree.

Expected Behavior

Implementations must guarantee the following behavior in addition to the global invariants:

If the provided lineage conflicts with an already-existing lineage in the backend, it must return BackendError::DuplicateLineage.

fn update_tree( &mut self, lineage: LineageId, new_version: VersionId, updates: SmtUpdateBatch, ) -> Result<MutationSet<SMT_DEPTH, Word, Word>, BackendError>

Performs the provided updates on the tree with the specified lineage, returning the mutation set that will revert the changes made to the tree.

Expected Behavior

Implementations must guarantee the following behavior in addition to the global invariants:

At most one new root must be added to the forest for the entire batch.

If applying the provided updates results in no changes to the tree, no new tree must be allocated.

fn update_forest( &mut self, new_version: VersionId, updates: SmtForestUpdateBatch, ) -> Result<Vec<(LineageId, MutationSet<SMT_DEPTH, Word, Word>)>, BackendError>

Performs the provided updates on the forest, setting all new tree states to have the provided new_version and returning a vector of the mutation sets that reverse the changes to each changed tree.

Expected Behavior

Implementations must guarantee the following behavior in addition to the global invariants:

At most one new root must be added to the forest for each target root in the provided updates.

If applying the provided updates results in no changes to a given lineage of trees in the forest, then no new tree must be allocated in that lineage.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.