miden_crypto::stark

Module air

Structs

FilteredAirBuilder

A wrapper around an AirBuilder that enforces constraints only when a specified condition is met.

FilteredMidenAirBuilder

A wrapper around a MidenAirBuilder that enforces constraints only when a specified condition is met.

VirtualPairCol

An affine linear combination of columns in a PAIR (Preprocessed AIR).

Enums

PairCol

A reference to a column in a PAIR (Preprocessed AIR).

Traits

Air

An algebraic intermediate representation (AIR) definition.

AirBuilder

A builder which contains both a trace on which AIR constraints can be evaluated as well as a method of accumulating the AIR constraint evaluations.

AirBuilderWithPublicValues

Extension trait for AirBuilder providing access to public values.

BaseAir

The underlying structure of an AIR.

BaseAirWithAuxTrace

An extension of BaseAir that includes support for auxiliary traces.

BaseAirWithPublicValues

An extension of BaseAir that includes support for public values.

ExtensionBuilder

Extension of AirBuilder for working over extension fields.

MidenAir

Super trait for all AIR definitions in the Miden VM ecosystem.

MidenAirBuilder

Super trait for all AIR builders in the Miden VM ecosystem.

PairBuilder

Trait for AirBuilder variants that include preprocessed data columns.

PermutationAirBuilder

Trait for builders supporting permutation arguments (e.g., for lookup constraints).