miden_crypto::stark::challenger

Trait CanObserve

pub trait CanObserve<T> {
    // Required method
    fn observe(&mut self, value: T);

    // Provided method
    fn observe_slice(&mut self, values: &[T])
       where T: Clone { ... }
}

A generic trait for absorbing elements into the transcript.

Absorbed elements update the internal sponge state, preparing it to deterministically produce future challenges.

Required Methods

fn observe(&mut self, value: T)

Absorb a single value into the transcript.

Provided Methods

fn observe_slice(&mut self, values: &[T])
where T: Clone,

Absorb a slice of values into the transcript.